requirements.txt
pandas>=1.5.3
numpy>=1.24.3
requests>=2.31.0
python-dotenv>=1.0.0
aliyun-python-sdk-core>=2.13.36
aliyun-python-sdk-dashscope>=1.0.2
tushare>=1.2.89
fastapi>=0.100.0
uvicorn>=0.22.0
pydantic>=2.0.3
python-jose>=3.3.0
passlib>=1.7.4
python-multipart>=0.0.6
beanie>=1.20.0
motor>=3.2.0